shifted util code into encoder.py
